requireMode: add require('@agoric/harden')
This yields the same as the previous 'def' global (which has been removed),
which is the one built-in to SES. We are not using the external
harden/makeHardener code yet.

refs #51
